SitePlaymarket.md
Google Play, also known as the Play Market or the Play Store - all this is the same program, which is one of the main ones in the Android operating system, because it opens the door to the world of applications and games for our smartphone and tablet. But sometimes it happens that when you launch the Google Play market, instead of the usual applications, we see a page with the inscription "Check the connection and try again" and the Retry button or the market gives some errors. Unfortunately, the Google Play application, like absolutely any other program, is not immune from all sorts of errors and failures in its work. This happens quite often, and if suddenly you have a similar inscription or the application does not load at all, then this article is for you.

In this article, we will try to answer the question: "What should I do if the Google Play market does not work or does not start on a smartphone or tablet?", As well as describe all kinds of problems when working with the Play Store and how to solve them.